requirements.txt
pyqt5
fpdf2